[ITERATION-1] add postgres, liquibase, transactions support; add posts table; update post repo (sql instead map stub);

+++ src/main/java/edu/yandex/project/repository/PostRepository.java
@@ -9,5 +9,5 @@ public interface PostRepository {
 
     List<PostEntity> findAll(@NonNull String titlePart, int offset, int limit);
 
-    int getPostCount();
+    Long getPostCount();
 }
